cold-email-assistant/requirements.txt
streamlit
openai
python-dotenv
fitz

